chore: model updates of offering and moving
