+++ src/main/kotlin/eda/shoppingBasket/service/eventing/shoppingBasket/SBOperation.kt
@@ -1,4 +1,4 @@
-package eda.shoppingBasket.service.eventing
+package eda.shoppingBasket.service.eventing.shoppingBasket
 
 enum class SBOperation {
     CREATED, UPDATED, DELETED, CHECKOUT
