+++ src/main/kotlin/eda/shoppingBasket/service/model/OfferingMapper.kt
@@ -5,29 +5,22 @@ import eda.shoppingBasket.service.model.dto.OfferingDTO
 import eda.shoppingBasket.service.model.entity.Offering
 
 class OfferingMapper {
-    fun toEntity(offeringDTO: OfferingDTO): Offering{
-        return Offering(
-            offeringID = offeringDTO.id,
-            quantity = offeringDTO.quantity,
-            price = offeringDTO.price,
-        )
-    }
-
-    fun toEntity(offeringEvent: OfferingEvent): Offering{
-        return Offering(
-            offeringID = offeringEvent.id,
-            quantity = offeringEvent.quantity,
-            price = offeringEvent.price,
-        )
-    }
-
-    fun toDto(offering: Offering): OfferingDTO{
+    fun toDto(offering:Offering): OfferingDTO{
         return OfferingDTO(
-            id = offering.offeringID,
+            id = offering.id,
+            status = offering.status,
             quantity = offering.quantity,
             price = offering.price,
-            productID = null,
-            status = offering.status
+            productID = null
+        )
+    }
+
+    fun toEntity(offeringDTO: OfferingDTO): Offering{
+        return Offering(
+            id = offeringDTO.id,
+            status = offeringDTO.status!!,
+            quantity = offeringDTO.quantity,
+            price = offeringDTO.price
         )
     }
 }
