+++ src/main/kotlin/eda/shoppingBasket/service/model/dto/OfferingInBasketDTO.kt
@@ -5,4 +5,4 @@ import java.util.UUID
 data class OfferingInBasketDTO(
     val offeringID: UUID,
     val quantity: Int
-)
+): DTO
