+++ src/main/kotlin/eda/shoppingBasket/service/model/entity/Offering.kt
@@ -7,34 +7,15 @@ import java.util.*
 @Entity
 class Offering(
     val quantity: Int,
-    val price: Float,
+    val price: Double,
     val status: Status = Status.ACTIVE,
     @Id
-    val offeringID: UUID
-
-) {
+    override val id: UUID = UUID.randomUUID()
+): AbstractEntity() {
     enum class Status {
         ACTIVE, INACTIVE, RETIRED;
         override fun toString(): String {
             return name.lowercase()
         }
     }
-    override fun equals(other: Any?): Boolean {
-        if (this === other) return true
-        if (javaClass != other?.javaClass) return false
-
-        other as Offering
-
-        return offeringID == other.offeringID
-    }
-
-    override fun hashCode(): Int {
-        var result = quantity
-        result = 31 * result + price.hashCode()
-        result = 31 * result + offeringID.hashCode()
-        return result
-    }
-//    i've added a ManyToOne in ShoppingBasketItem instead of a OneToMany reference here
-//    @OneToMany
-//    val shoppingBasketItems: MutableList<ShoppingBasketItem> = mutableListOf()
 }
